day8

+++ bbs/index.php
@@ -1,3 +1,35 @@
+<?php
+session_start();
+require('library.php');
+
+if (isset($_SESSION['id']) && isset($_SESSION['name'])) {
+    $id = $_SESSION['id'];
+    $name = $_SESSION['name'];
+} else {
+    header('Location: login.php');
+    exit();
+}
+
+$db = dbconnect();
+
+// メッセージの投稿
+if ($_SERVER['REQUEST_METHOD'] === 'POST') {
+    $message = filter_input(INPUT_POST, 'message', FILTER_SANITIZE_STRING);
+    $stmt = $db->prepare('insert into posts (message, member_id) values(?,?)');
+    if (!$stmt) {
+        die($db->error);
+    }
+
+    $stmt->bind_param('si', $message, $id);
+    $success = $stmt->execute();
+    if (!$success) {
+        die($db->error);
+    }
+
+    header('Location: index.php');
+    exit();
+}
+?>
 <!DOCTYPE html>
 <html lang="ja">
 
@@ -19,7 +51,7 @@
         <div style="text-align: right"><a href="logout.php">ログアウト</a></div>
         <form action="" method="post">
             <dl>
-                <dt>○○さん、メッセージをどうぞ</dt>
+                <dt><?php echo h($name); ?>さん、メッセージをどうぞ</dt>
                 <dd>
                     <textarea name="message" cols="50" rows="5"></textarea>
                 </dd>
@@ -31,13 +63,28 @@
             </div>
         </form>
 
+        <?php $stmt = $db->prepare('select p.id, p.member_id, p.message, p.crated, m.name, m.picture from poste p, members m where m.id=p.member_id order by id desc');
+        if (!$stmt) {
+            die($db->error);
+        }
+        $success = $stmt->execute();
+        if (!$success) {
+            die($db->error);
+        }
+
+        $stmt->bind_result($id, $member_id, $message, $created, $name, $picture);
+        while ($stmt->fetch());
+        ?>
         <div class="msg">
-            <img src="member_picture/" width="48" height="48" alt=""/>
-            <p>○○<span class="name">（○○）</span></p>
-            <p class="day"><a href="view.php?id=">2021/01/01 00:00:00</a>
+            <?php if ($picture): ?>
+              <img src="member_picture/<?php echo h($picture); ?>" width="48" height="48" alt=""/>
+            <? endif; ?>
+            <p><?php echo h($message); ?><span class="name">（<?php echo h($name); ?></span></p>
+            <p class="day"><a href="view.php?id=<?php echo h($id); ?>"><?php echo h($created); ?></a>
                 [<a href="delete.php?id=" style="color: #F33;">削除</a>]
             </p>
         </div>
+        <? endwhile; ?>
     </div>
 </div>
 </body>
